refactor: extract MessageRow component from RoomView
