+++ app/views/RoomView/components/index.ts
@@ -1,3 +1,4 @@
 export * from './EncryptedRoom';
 export * from './HeaderCallButton';
+export * from './MessageRow';
 export * from './MissingRoomE2EEKey';
